Build Bosnia & Herzegovina using adm1_codes
https://github.com/tmrowco/electricitymap/issues/951

In the recent natural earth update the Bosnia & Herzegovina (BIH)
country shape is missing most of the eastern part of the country for
some unknown reason.  This causes BIH to be displayed incorrectly on the
map.  Using adm1_codes from the natural earth states file to build the
area fixes the problem.