Add US (MISO) (#1024)
* Create US-MISO production parser

* Complete US-MISO parser

* Update readme with US-MISO info

* Create US-MISO zone

Disabling states of which MISO is roughly composed. Eventually this will
be replaced by a correct shapefile.

* Fix bug in update_state that prevents production parsers running

If one zone was passed as an argument this resulted in the zone
name being assigned as the first letter only. Also now that the
script can take 'all' as an argument the looping logic assumed
that 'zone_config' was the entire file but for a single zone
only a small portion of the file was assigned. This resulted in
a keyerror with no error message displayed.

The solution is not to use list() to create zone_name and to
ensure the whole zone file is assigned to zone_config. Also
provide a message if zone has no production parser.

* Update mockserver state

* Change USA formatting and remove pyiso reference