Separate Russia into regions
The data source used by RU.py does not include some parts
of the country (mainly north-west and eastern). Using states
allows a better approximation. It should be noted that
RUS-2603 is included despite only the lower 1/3 being part
of the data source. This is because of it's large generation
capacity (~7GW).

Kaliningrad Oblast is again separated from the main RU zones.