Improve accuracy of Russian zones (#1092)
* Get Russia from natural earth states file

* Separate Russia into regions

The data source used by RU.py does not include some parts
of the country (mainly north-west and eastern). Using states
allows a better approximation. It should be noted that
RUS-2603 is included despite only the lower 1/3 being part
of the data source. This is because of it's large generation
capacity (~7GW).

Kaliningrad Oblast is again separated from the main RU zones.

* Create new RU-E zone

RU-E (mainly eastern Russia) approximates the area which we
don't have generation data for.

* Generate new world.json