Albania export's emission factor
Default source for generation in 2015:
https://www.iea.org/statistics/statisticssearch/report/?country=ALBANIA&product=electricityandheat&year=2015

New hydro power is supposed to come online this year. Currently up to 1800 MW generation based on the links I gave in #1062

Bit longer article on the current and future capacity development, if you're interested:
https://www.energyworldmag.com/electricity-generation-capacity-development-albania/
It says there is a CCGT gas power plant, but according to other [news](https://seenews.com/news/albania-plans-to-feed-gas-from-tap-pipeline-into-idled-vlora-tpp-532173), it was never put into operation since 2011 due to mechanical faults. 

Also there is some generation from waste but with max of 2.85 MW, so insignificant atm. 
PV capacity will slowly rise, tenders are expected. If any new capacity is added, it will still be represented by the "100% renewable" ratio.
Potential is between 1500 to 1800 kWh/kWp, so 100 MWp PV will translate into 150 to 180 GWh/a in the future.