Alter NA exchange regex
NA exchanges take the form exchange_name:(whitespace)(value)
The NA->ZM exchange was failing to be read on certain values.
The cause of this was the large amount of whitespace between
: and (value), which caused the colon to be read as an
alphabetic character.

This fix makes the regex match any character except digits
and minus sign in the first group, the second group matches
the generation value.