SG: don't fail when unable to OCR
Solar production is a very small part of Singapore electricity, so when
it can't be retrieved, just log warning and move on.

Add the OCR'd text to log message to possibly aid debugging.

Currently OCR is stuck on parsing the datetime "2018-03-31 11:44"
which it reads as "2018-03-31 1 1 :44". Maybe the regexp could be
improved to handle that, but in meanwhile, log better.