Use UTC for mocking the future data timestamp (#2032)
When determining if a data timestamp is reasonable, it is compared to
the current UTC time. If someone is runing the parser tests in a
timezone other than UTC, the mock future time needs to be created using
the current UTC time.

This commit mocks parser data timestamps relative to the current UTC
time.