Add parser for Gulf Cooperation Council countries (#2360)
* Add parser for Gulf Cooperation Council countries

Which are the United Arab Emirates, Bahrain, Saudi Arabia, Oman, Qatar, and
Kuwait.

Oman and Saudi Arabia are currently not returning data. If/when data is
available again, the parser will be able to retrieve it.

* README: put Kuwait sources on separate lines

* Add zones for AE, BH, OM, QA and SA

* GCCIA: Rename fetch_production to fetch_consumption

Wasn't sure if it'd work with only fetch_consumption. Changing it as per reviewer suggestion


Co-authored-by: q-- <io7io@ya.ru>