Update of Spain's generating capacity (#2432)
Spain is starting its version of the german Energiewende, in which solar capacity is growing at a high rate, so although data was from november 2019, it was already outdated.

The source of the data is REE's monthly report on generating capacity of March 2020. I added plus 341 MW of a solar PV park installed at the beggining of this April. I myself will review again installed capacities when May's report is uploaded, so that summer data is updated through all three summer months.

Co-authored-by: Roberta Hunt <roberta.hunt@tmrow.com>