Fix stale data check in Singapore solar parser.
Use .total_seconds() instead of .seconds. The former is the total
duration of the timedelta in seconds. The latter is just the seconds
part of a timedelta which can actually cause false positives and
false negatives. For example timedelta(days=1, seconds=0) would pass
when it should fail, and timedelta(days=-1, seconds=86398) would fail
when it should pass (e.g. the image was just generated but the local
clock is a slightly behind the remote clock).