Panama: report fossil plants with known fuel type as coal/oil/gas (#2642)
* Add parser for Gulf Cooperation Council countries

Which are the United Arab Emirates, Bahrain, Saudi Arabia, Oman, Qatar, and
Kuwait.

Oman and Saudi Arabia are currently not returning data. If/when data is
available again, the parser will be able to retrieve it.

* README: put Kuwait sources on separate lines

* Add zones for AE, BH, OM, QA and SA

* GCCIA: Rename fetch_production to fetch_consumption

Wasn't sure if it'd work with only fetch_consumption. Changing it as per reviewer suggestion

Co-Authored-By: Roberta Hunt <roberta.hunt@tmrow.com>

* Panama: report known coal plants as 'coal'

Reports two known coal plants as 'coal' after extracting their
generation from the generation by unit breakdown on the page

* Report known Oil-fired plant as "oil"

Also removed leftover print statement

* Report Costra Norte  generation as 'gas'

* Panama: Take care of negative production

* Add name to contributors for PA

Co-authored-by: q-- <io7io@ya.ru>