Add Himachal Pradesh (India) parser (#2686) (#2654)
* Add draft Himachal Pradesh (India) parser (#304)

This uses the real time data page from the Himachal Pradesh SLDC:
https://hpsldc.com/intra-state-power-transaction/

This currently does production only.

* Improve draft IN-HP parser based on comments

* Note: this still only includes fetch_production().
* Changed fuzzy plant name matching to exact name matching.
* Some other small fixes.

* Add IN-HP to zones.json

See the comment on the PLANT_NAMES_TO_TYPES dict for how the capacity
was calculated.

I used the generate-zone-bounding-boxes script to calculate the bounding
box.

* Add unit test for IN-HP fetch_production()

This replaces the testing code in the parser. Also did some final
cleanup and comments. Therefore this parser is now ready to go from
my perspective.

* Add stubs for unimplemented IN-HP functions with explanation

This is to provide a starting point if we decide to implement these in future.

Co-authored-by: Chris <30196510+systemcatch@users.noreply.github.com>