Add Puerto Rico (#2531)
* Add parser for Gulf Cooperation Council countries

Which are the United Arab Emirates, Bahrain, Saudi Arabia, Oman, Qatar, and
Kuwait.

Oman and Saudi Arabia are currently not returning data. If/when data is
available again, the parser will be able to retrieve it.

* README: put Kuwait sources on separate lines

* Add zones for AE, BH, OM, QA and SA

* GCCIA: Rename fetch_production to fetch_consumption

Wasn't sure if it'd work with only fetch_consumption. Changing it as per reviewer suggestion

Co-Authored-By: Roberta Hunt <roberta.hunt@tmrow.com>

* Add Puerto Rico

See #2531

Gives the generation breakdown (updated every 30 minutes)
Can't parse production forecast yet

* Puerto Rico: get rid of hardcoded zone key

* Static carbon intensity if no renewables breakdown

Switches to a static carbon intensity based on the renewable output of
the last twelve months (to account for seasonal fluctuations in wind,
solar and hydro output) when the renewables breakdown page is not
available

* Changed name of zone & generated bounding box

* Generated world.json again

* Update flag & translations

* Renamed to 'PR' again

Co-authored-by: q-- <io7io@ya.ru>
Co-authored-by: Roberta Hunt <roberta.hunt@tmrow.com>
Co-authored-by: Felix Qvist <felix.qvist@gmail.com>