AUS-TAS-KI: add more logging
There currently is an issue where sometimes the 'technologies' key isn't
present on the object received from SignalR, and because debug-level
messages aren't logged to Kibana it's unclear why.

This commit will make it log the full object as an error so it shows up
in Kibana and can be examined.