Change ENTSOE parse_production so it includes all storage as consumption
ef7219d made parse_production include hydro storage as consumption, but it was
suggested:

https://github.com/tmrowco/electricitymap-contrib/pull/2805#discussion_r526996343

that we should include all storage in case we add other types in future.
At the moment hydro storage is the only one there is, but this code means any future
storage params will be handled as well.