Make ENTSOE fetch_consumption() include self-consumption (fix #2710) (#2814)
* Make ENTSOE fetch_consumption() include self-consumption (fix #2710)

This adds power plant self-consumption (which is found in the production data)
to the total consumption value for a given timezone.

It has a limitation in that consumption data is often available a few minutes before
production data, so self-consumption will not be included in total consumption for the
most recent timezone. E.g. running the parser:

python test_parser.py NL consumption

at 17:32 UTC returns consumption data for the most recent datetime of 16:45 (total
value 14780), but does not include self-consumption as at this point production data is only
available until 16:30. However if you run it again at 17:40, production data for 16:45 is
now available, so the total consumption for the most recent timezone of 16:45 is now 15611.
If this happens it will be logged as a warning.

Co-authored-by: Felix Qvist <felix.qvist@gmail.com>