Fix SV parser
The parser for El Salvador is broken. ut.com.sv now [reports wind
generation](http://estadistico.ut.com.sv/OperacionDiaria.aspx), and has
done this by inserting it in the output as the second element, after
biomass but before everything else.

As the result of this, every reported generation other than biomass at
ElectricityMap is wrong because production values are mapped to the
wrong type.

The fact that there is still some output on ElectricityMap is because
when there's no solar output at night, it is absent from the object so
the number of keys appears to be correct during that time. But when day
dawns, the object has "solar" data so there is one more key than the
parser is programmed to handle and it errors.

This commit fixes things by mapping wind (and consequently everything
else) correctly.