Update to CO2 values for Slovenia
I've added a custom value for coal in Slovenia. Only full time coal power plant in Slovenia is TEŠ6. Official CO2 value for is TEŠ6 869g/KW. When TEŠ6 is out of service, old TEŠ5 power plant is turned on but that is rare and an exception. I believe that CO2 of TEŠ6 most accurately represents coal CO2 value.