Ignore flaky code push errors (#3046)
Silent failing is not nice. But the code push job is flaky and blocks
our deployment pipeline... We might be able to stabilize this by upgrading
to https://github.com/microsoft/appcenter-cli (note that the current version
of the code push CLI is 3 years old).

For now we should simply monitor that code is being shipped to the Android
and iOS apps.