update USA capacites Q1
I updated the capacities of the USA with the same source as before. EIA publishes the new data on a monthly basis with a 3 month delay. I'll do an update every quarter since it is a lot of work going through all the balacing authorities and most of the USA isn't colored at the moment anyway. 

Also I added Puerto Ricos capacities. They're included in the same report.