Added CA-SK current consumption
The following page shows a System Load widget in megawatts:
https://www.saskpower.com/Our-Power-Future/Our-Electricity/Electrical-System/System-Map

It makes an AJAX call to:
https://www.saskpower.com/ignitionapi/Content/GetNetLoad

That endpoint simply returns a number as a string. I've interpreted
this to be the current consumption based solely on the website's
language surrounding this number, "system load" and "net load".

fetch_consumption(...) has been implemented accordingly.