Assign SG image crop dimensions to variables
The crop function expects Tuple arguments that are
integers. Though they were x.0 float values, the
previous implementation was a tuple of floats. I've
explicitly converted them with int(...) in addition
to giving them variable names for clarity.

Additionally, I shifted the left pixel dimension further
in from ~52% to ~65% from the left of the image. This
results in a smaller crop for tesseract to process
later.