Update NL wind/solar based on 2023-06-02 CBS data, link to main CBS table in DATA_SOURCES (#5450)
* Update NL wind/solar based on 2023-06-02 CBS data

I tried to find the current capacity numbers in the listed sources, but couldn't. The already-cited CBS source published an update to the table of installed capacities as of 2023-06-02, see here: https://opendata.cbs.nl/#/CBS/nl/dataset/82610NED. The numbers I'm including are the end-of-year ones for 2022. As far as I know there are no more up-to-date numbers than this, but happy to be pointed to them if they do exist.

* Update NL wind/solar source to source CBS table

This table gets updated, whereas the previous source linked to a specific CBS press release that does not get updated.

* Fix CBS link to Dutch table (English-language table does not include solar)

---------

Co-authored-by: Mathilde Daugy <107848894+mathilde-daugy@users.noreply.github.com>