Moving IN=WE to a new data source which has 24 hour datetime data unlike 12 hour
The older API endpoints provided data in 12 hour format without AM and PM, because of which there was quite a lot of logic in code to parse that data.

The new APIs provide data in 24 hour format which is more correct and simple to process.

Also made changes to methods to use Pandas groupby on datetime hour which simplified the code even further and fixed some bugs which would arise in the earlier version for missing datapoints from the source