refactor/fix(parser): swap arrow in favour of datetime in CO, etc.
This commit replaces arrow, does some additional refactoring, and fixes
a small bug.

The bug was that the requested data was from a date determined by a
parser executing local datetime instead of the API local datetime.

The additional refactoring includes for example:
- De-nesting code under conditionals by raising an error early if
  needed under a conditional, but then de-nesting the other code block.
- Renaming local functions by prefixing them with an underscore
- Coalescing code that sometimes is run once and sometimes looped over
  by always looping (but sometimes only one iteration).

This commit was made without the ability to test the parser against the
API, as the API was down or wasn't accessible from Sweden.