fix(locales): make sv.json's net exchange translation consistent
The swedish translation of "net exchange" occur in two places, where
the initial one was translated as "nettoöverföring", while the one
added later was translated as "nettohandel".

My take is that the initial translation of "nettoöverföring" better
conveys that this chart is about net import/export than "nettohandel"
does. No matter what is decided, I think we should make the
translations consistent to be one way or the other.