refactor: replace risky strategy of distinguishing AM from PM hours
The drop duplicates strategy relied on both AM and PM data included
exactly 1:1 entries minute entries, but the data indicated that
sometimes a minute was missed or similar, so this was unreliable. The
strategy would result in a few minutes from a AM hour showing up in a PM
hour or vice versa.

This new strategy relies on the 12:xx AM hours are shown first, followed
by 01-11:xx hours, followed by 12:xx PM hours, and is a robuster
robuster assumption that doesn't need us to zero the seconds etc.