Package config separately
We can ensure that the config folder is installed to consuming projects by adding an `__init__.py` file and adding it to the `packages` directive. This will mean that it will be installed simply as `config` in the `/.venv/lib/python/site-packages/` directory of the consuming project and can then be accessed using a relative path from the `electricitymap-config` package. This is a good deal than other solutions proposed, but one disadvantage is that the new `config` package has a very undescriptive name. It could be possible to change this, but this would involve updating the callers which makes this a bigger change. As such, I suggest going with this solution as the other possibilities are more difficult.

How the files look when installed:

```
ls .venv/lib/python3.10/site-packages/config
__init__.py   data_centers  defaults.yaml Earthfile     exchanges     retired_zones zones
```

Verifying that the constants work:
```
poetry run python
Python 3.10.13 (main, May  1 2025, 13:41:41) [Clang 17.0.0 (clang-1700.0.13.3)] on darwin
Type "help", "copyright", "credits" or "license" for more information.
>>> from electricitymap.contrib.config import ZONES_CONFIG, EXCHANGES_CONFIG
>>> type(ZONES_CONFIG)
<class 'dict'>
```