Move parsers into package
The parsers need to be in the electricitymap/contrib directory if they are going to be used by upstream projects. The current setup uses separate directory AND a symlink with imports using both paths. This breaks when using the parsers in a package, so we need to use a single approach.